refactor: でんしょうおのページ責務を分離
