+++ src/app/denshouo/DenshouoScene.tsx
@@ -1,0 +1,779 @@
+'use client';
+
+import { motion, useReducedMotion } from 'framer-motion';
+import { Canvas, useFrame } from '@react-three/fiber';
+import { useEffect, useMemo, useRef, useState } from 'react';
+import * as THREE from 'three';
+import type { WeatherType } from '@/store';
+import { waveVertexShader, waveFragmentShader } from '@/shaders/wave';
+
+const bubbleSpecs = [
+    { left: '8%', top: '18%', size: 16, delay: 0.0, duration: 8.0 },
+    { left: '18%', top: '68%', size: 10, delay: 1.2, duration: 7.2 },
+    { left: '31%', top: '44%', size: 22, delay: 0.6, duration: 9.5 },
+    { left: '52%', top: '72%', size: 14, delay: 1.8, duration: 8.8 },
+    { left: '66%', top: '22%', size: 18, delay: 0.3, duration: 10.0 },
+    { left: '81%', top: '56%', size: 12, delay: 1.0, duration: 7.8 },
+    { left: '90%', top: '28%', size: 24, delay: 2.0, duration: 11.0 },
+];
+
+const fishSpecs = [
+    { top: '20%', left: '12%', width: 72, delay: 0.0, duration: 18 },
+    { top: '58%', left: '68%', width: 56, delay: 1.4, duration: 16 },
+    { top: '74%', left: '24%', width: 42, delay: 2.1, duration: 14 },
+];
+
+const rainLightSpecs = [
+    { left: '18%', width: 6, rotate: -9, delay: 0.0, duration: 9.5 },
+    { left: '43%', width: 9, rotate: 5, delay: 1.8, duration: 11.0 },
+    { left: '70%', width: 7, rotate: -4, delay: 0.9, duration: 10.2 },
+];
+
+const marineSnowSpecs = Array.from({ length: 34 }, (_, index) => ({
+    left: 5 + ((index * 37) % 90),
+    size: 2 + ((index * 17) % 5),
+    drift: -24 + ((index * 29) % 49),
+    duration: 10 + ((index * 13) % 11),
+    delay: -((index * 7) % 16),
+    opacity: 0.20 + ((index * 11) % 25) / 100,
+    blur: index % 4 === 0 ? 2 : 0,
+}));
+
+function detectWebGLSupport() {
+    if (typeof document === 'undefined') return false;
+
+    const canvas = document.createElement('canvas');
+    try {
+        return Boolean(
+            canvas.getContext('webgl2', { failIfMajorPerformanceCaveat: false })
+                ?? canvas.getContext('webgl', { failIfMajorPerformanceCaveat: false })
+        );
+    } catch {
+        return false;
+    }
+}
+
+function useWebGLSupport() {
+    const [supported, setSupported] = useState(false);
+
+    useEffect(() => {
+        // WebGL support is an external browser capability; read it after mount.
+        // eslint-disable-next-line react-hooks/set-state-in-effect
+        setSupported(detectWebGLSupport());
+    }, []);
+
+    return supported;
+}
+
+const OCEAN_BUBBLE_DATA = (() => {
+    const count = 100;
+    const positions = new Float32Array(count * 3);
+    const scales = new Float32Array(count);
+    const speeds = new Float32Array(count);
+    const seeds = new Float32Array(count);
+
+    for (let index = 0; index < count; index++) {
+        positions[index * 3] = (Math.random() - 0.5) * 20;
+        positions[index * 3 + 1] = (Math.random() - 0.5) * 14;
+        positions[index * 3 + 2] = -Math.random() * 12 - 1.2;
+        scales[index] = Math.random() * 2.0 + 0.8;
+        speeds[index] = Math.random() * 0.9 + 0.5;
+        seeds[index] = Math.random();
+    }
+
+    return { positions, scales, speeds, seeds };
+})();
+
+// ── Animated swimming fish cursor ─────────────────────────────────────────────
+// Draws a procedurally animated fish on a canvas overlay.
+// The fish body bends with a travelling sine wave, tail fans,
+// and the whole fish rotates toward the velocity vector.
+function FishCursor() {
+    const canvasRef = useRef<HTMLCanvasElement>(null);
+    const stateRef  = useRef({
+        // current display position (smoothed toward raw)
+        x: -200, y: -200,
+        // raw mouse position
+        rawX: -200, rawY: -200,
+        // velocity (pixels/frame, exponential moving average)
+        vx: 0, vy: 0,
+        // true = fish faces left (horizontal mirror via scale(-1,1))
+        facingLeft: false,
+        // pitch angle: up/down tilt in radians (capped ±π/2, always right-side-up)
+        pitchAngle: 0,
+        // swim phase (drives sine wave)
+        phase: 0,
+        // mouth open (click)
+        mouthOpen: false,
+        // is fine pointer
+        fine: false,
+        // raf id
+        raf: 0,
+    });
+
+    useEffect(() =>
+    {
+        const s = stateRef.current;
+        s.fine = window.matchMedia('(pointer: fine)').matches;
+        if (!s.fine) return;
+
+        const onMove = (e: PointerEvent) => {
+            if (e.pointerType !== 'mouse') return;
+            s.rawX = e.clientX;
+            s.rawY = e.clientY;
+        };
+        const onDown = (e: PointerEvent) => {
+            if (e.pointerType !== 'mouse') return;
+            s.mouthOpen = true;
+        };
+        const onUp = (e: PointerEvent) => {
+            if (e.pointerType !== 'mouse') return;
+            s.mouthOpen = false;
+        };
+        window.addEventListener('pointermove',   onMove,  { passive: true });
+        window.addEventListener('pointerdown',   onDown,  { passive: true });
+        window.addEventListener('pointerup',     onUp,    { passive: true });
+        window.addEventListener('pointercancel', onUp,    { passive: true });
+        return () => {
+            window.removeEventListener('pointermove',   onMove);
+            window.removeEventListener('pointerdown',   onDown);
+            window.removeEventListener('pointerup',     onUp);
+            window.removeEventListener('pointercancel', onUp);
+        };
+    }, []);
+
+    useEffect(() => {
+        const canvas = canvasRef.current;
+        if (!canvas) return;
+        const s = stateRef.current;
+        if (!window.matchMedia('(pointer: fine)').matches) return;
+
+        const resize = () => {
+            canvas.width  = window.innerWidth  * window.devicePixelRatio;
+            canvas.height = window.innerHeight * window.devicePixelRatio;
+            canvas.style.width  = `${window.innerWidth}px`;
+            canvas.style.height = `${window.innerHeight}px`;
+        };
+        resize();
+        window.addEventListener('resize', resize);
+
+        const ctx = canvas.getContext('2d')!;
+        const dpr = window.devicePixelRatio;
+
+        // ── Drawing helpers ──────────────────────────────────────────────────
+        const drawClownfish = (t: number, mouthOpen: boolean) => {
+            const SCALE = 1.45; const S = SCALE;
+            const swingAmp = Math.max(0.18, Math.min(0.9, Math.hypot(s.vx, s.vy) * 0.04));
+            const tailSwing = Math.sin(t * 7.0 + s.phase) * swingAmp;
+            const bodyBend = Math.sin(t * 7.0 + s.phase + 0.8) * swingAmp * 0.35;
+
+            ctx.save();
+            const tailAngle = tailSwing * 1.4; const tailLength = 12.1 * S; const tailRoot = -10.2 * S;
+            ctx.save();
+            ctx.rotate(bodyBend * 0.6);
+            ctx.beginPath(); ctx.moveTo(tailRoot, 0);
+            ctx.lineTo(tailRoot - Math.cos(tailAngle + 0.5) * tailLength, Math.sin(tailAngle + 0.5) * tailLength * 1.15);
+            ctx.lineTo(tailRoot - Math.cos(tailAngle - 0.5) * tailLength, Math.sin(tailAngle - 0.5) * tailLength * 1.15);
+            ctx.closePath();
+            const tailGrad = ctx.createLinearGradient(tailRoot, 0, tailRoot - tailLength, 0);
+            tailGrad.addColorStop(0, 'rgba(180,80,0,0.95)');
+            tailGrad.addColorStop(1, 'rgba(140,55,0,0.6)');
+            ctx.fillStyle = tailGrad; ctx.fill(); ctx.strokeStyle = 'rgba(20,8,0,0.7)'; ctx.lineWidth = 0.9; ctx.stroke();
+            ctx.restore();
+
+            ctx.save(); ctx.rotate(bodyBend);
+            const bodyGrad = ctx.createRadialGradient(-2*S, -4*S, 1, 1*S, 0, 14*S);
+            bodyGrad.addColorStop(0, '#FFB347'); bodyGrad.addColorStop(0.55, '#FF8C1A'); bodyGrad.addColorStop(1, '#C46200');
+            ctx.beginPath();
+            ctx.moveTo(-11.2 * S, 0);
+            ctx.bezierCurveTo(-8.2 * S, -9.2 * S, 4.2 * S, -11.2 * S, 11.3 * S, -4.0 * S);
+            ctx.quadraticCurveTo(14.0 * S, -0.9 * S, 12.3 * S, 3.4 * S);
+            ctx.bezierCurveTo(6.0 * S, 11.4 * S, -7.4 * S, 9.8 * S, -11.2 * S, 0);
+            ctx.fillStyle = bodyGrad; ctx.fill(); ctx.strokeStyle = 'rgba(35,13,2,0.5)'; ctx.lineWidth = 1.15; ctx.stroke();
+
+            ctx.save(); ctx.clip();
+            ctx.fillStyle = 'rgba(255,255,255,0.86)';
+            ctx.beginPath();
+            ctx.moveTo(-4.4 * S, -8.3 * S);
+            ctx.bezierCurveTo(-2.4 * S, -9.4 * S, -0.5 * S, -8.6 * S, 0.3 * S, -6.6 * S);
+            ctx.bezierCurveTo(-1.0 * S, -2.4 * S, -0.9 * S, 2.6 * S, 0.6 * S, 7.0 * S);
+            ctx.bezierCurveTo(-1.6 * S, 8.5 * S, -3.8 * S, 7.8 * S, -4.8 * S, 5.4 * S);
+            ctx.bezierCurveTo(-3.7 * S, 1.8 * S, -3.6 * S, -3.9 * S, -4.4 * S, -8.3 * S);
+            ctx.fill();
+            ctx.fillStyle = 'rgba(255,255,255,0.74)';
+            ctx.beginPath();
+            ctx.moveTo(4.2 * S, -7.3 * S);
+            ctx.bezierCurveTo(5.6 * S, -7.6 * S, 7.0 * S, -6.8 * S, 7.5 * S, -5.3 * S);
+            ctx.bezierCurveTo(6.5 * S, -1.8 * S, 6.6 * S, 2.7 * S, 7.8 * S, 6.0 * S);
+            ctx.bezierCurveTo(6.5 * S, 7.2 * S, 4.9 * S, 6.8 * S, 4.2 * S, 5.0 * S);
+            ctx.bezierCurveTo(5.0 * S, 1.3 * S, 5.0 * S, -3.4 * S, 4.2 * S, -7.3 * S);
+            ctx.fill();
+            ctx.strokeStyle = 'rgba(40,17,4,0.24)';
+            ctx.lineWidth = 0.65;
+            ctx.beginPath();
+            ctx.moveTo(-5.2 * S, -7.6 * S); ctx.bezierCurveTo(-3.4 * S, -2.2 * S, -3.3 * S, 2.4 * S, -5.3 * S, 7.0 * S);
+            ctx.moveTo(3.5 * S, -6.7 * S); ctx.bezierCurveTo(5.0 * S, -2.0 * S, 5.1 * S, 2.4 * S, 3.8 * S, 6.2 * S);
+            ctx.stroke();
+            ctx.restore();
+
+            const eyeX = 7.8 * S; const eyeY = -2.8 * S;
+            ctx.beginPath(); ctx.arc(eyeX, eyeY, 3.0 * S, 0, Math.PI * 2); ctx.fillStyle = '#111'; ctx.fill();
+            ctx.beginPath(); ctx.arc(eyeX + 0.7*S, eyeY - 0.7*S, 0.8*S, 0, Math.PI * 2); ctx.fillStyle = 'white'; ctx.fill();
+
+            const mouthX = 11.2 * S; const mouthY = 1.5 * S;
+            if (mouthOpen) {
+                ctx.beginPath(); ctx.moveTo(mouthX, mouthY); ctx.lineTo(mouthX - 5*S, mouthY - 4*S); ctx.lineTo(mouthX - 5*S, mouthY + 4*S); ctx.closePath();
+                ctx.fillStyle = 'rgba(26,5,0,0.95)'; ctx.fill(); ctx.strokeStyle = 'rgba(13,5,0,0.8)'; ctx.lineWidth = 0.8; ctx.stroke();
+            } else {
+                ctx.beginPath();
+                ctx.quadraticCurveTo(11.4 * S, 2.1 * S, 10.2 * S, 2.6 * S);
+                ctx.strokeStyle = 'rgba(55,18,4,0.45)';
+                ctx.lineWidth = 0.7;
+                ctx.stroke();
+            }
+            ctx.restore(); ctx.restore();
+        };
+
+        const drawAnglerfish = (t: number, mouthOpen: boolean) => {
+            const SCALE = 1.35; const S = SCALE; // slightly smaller/chubbier scale
+            const swingAmp = Math.max(0.18, Math.min(0.9, Math.hypot(s.vx, s.vy) * 0.04));
+            const tailSwing = Math.sin(t * 8.0 + s.phase) * swingAmp;
+            const bodyBend = Math.sin(t * 8.0 + s.phase + 0.8) * swingAmp * 0.35;
+
+            ctx.save();
+            ctx.rotate(bodyBend * 0.3);
+
+            // Stubby cute deep-sea tail (scaled up)
+            ctx.beginPath();
+            ctx.moveTo(-10*S, 0); 
+            ctx.lineTo(-17*S, 6*S + tailSwing*10);
+            ctx.quadraticCurveTo(-20*S, 0, -17*S, -6*S + tailSwing*10);
+            ctx.fillStyle = '#1e1b4b'; // deep indigo tail
+            ctx.fill(); 
+            ctx.strokeStyle = '#020617'; ctx.lineWidth = 1.2; ctx.stroke();
+
+            // Pectoral fin flap (soft rounded)
+            ctx.beginPath(); ctx.moveTo(-3*S, 2*S);
+            const finSwing = Math.sin(t*10)*S;
+            ctx.quadraticCurveTo(0*S, 4*S+finSwing, -1*S, 7*S+finSwing); 
+            ctx.quadraticCurveTo(-4*S, 4*S, -3*S, 2*S);
+            ctx.fillStyle = '#312e81'; ctx.fill(); ctx.stroke();
+
+            // Bulbous, chubby potato body 
+            const bulbGrad = ctx.createRadialGradient(2*S, -2*S, 2*S, 1*S, -1*S, 12*S);
+            bulbGrad.addColorStop(0, '#4338ca'); // soft purple/indigo top
+            bulbGrad.addColorStop(1, '#1e1b4b'); // dark belly
+            
+            ctx.beginPath();
+            ctx.moveTo(-12*S, 0); 
+            // Plump back
+            ctx.bezierCurveTo(-6*S, -16*S, 8*S, -14*S, 13*S, -4*S); 
+            
+            // Mouth
+            if (mouthOpen) {
+                // Happy wide gasp
+                ctx.lineTo(8*S, 0*S); // deep smile inside
+                ctx.quadraticCurveTo(10*S, 14*S, 14*S, 9*S); // drop round jaw
+            } else {
+                // Cute round face with slight underbite
+                ctx.lineTo(13*S, 7*S);
+            }
+            
+            // Plump belly
+            ctx.bezierCurveTo(4*S, 16*S, -6*S, 10*S, -12*S, 0);
+            ctx.fillStyle = bulbGrad; ctx.fill(); 
+            ctx.strokeStyle = '#1e1b4b'; ctx.lineWidth = 1.6; ctx.stroke();
+
+            // Cute chunky snaggleteeth (scaled up as requested)
+            ctx.fillStyle = '#f8fafc';
+            ctx.beginPath();
+            // Just cute little rounded upper teeth (scaled up)
+            ctx.moveTo(13.5*S, -3*S); ctx.lineTo(11.5*S, 3.5*S); ctx.lineTo(9.5*S, -3*S); // scaled small fang
+            ctx.moveTo(10.5*S, -3*S); ctx.lineTo(7.5*S, 6.0*S);  ctx.lineTo(4.5*S, -4*S); // scaled big fang
+            
+            if (mouthOpen) {
+                // Cute lower teeth pointing up
+                ctx.moveTo(13*S, 8.5*S); ctx.lineTo(12*S, 4*S); ctx.lineTo(11*S, 8*S);
+                ctx.moveTo(10*S, 8*S);   ctx.lineTo(9.5*S, 3*S); ctx.lineTo(8*S, 7*S);
+                ctx.moveTo(8*S,  7*S);   ctx.lineTo(7*S, 4*S);  ctx.lineTo(6*S, 6*S);
+            } else {
+                // Just snaggle teeth poking out over the closed mouth!
+                ctx.moveTo(12*S, 7*S); ctx.lineTo(11*S, 2*S); ctx.lineTo(10*S, 6*S);
+                ctx.moveTo(9*S,  6*S); ctx.lineTo(8.5*S, 3*S); ctx.lineTo(7*S, 4.5*S);
+            }
+            ctx.fill();
+
+            // Pure, simple, classically cute flat eye
+            const eyeX = 7 * S;
+            const eyeY = -4.5 * S;
+            ctx.beginPath(); 
+            ctx.arc(eyeX, eyeY, 2.6 * S, 0, Math.PI * 2);
+            ctx.fillStyle = '#0f172a'; // solid dark base
+            ctx.fill();
+            
+            // Single large glowing cyan catchlight for an innocent look
+            ctx.beginPath(); 
+            ctx.arc(eyeX + 0.6 * S, eyeY - 0.6 * S, 1.2 * S, 0, Math.PI * 2);
+            ctx.fillStyle = '#cffafe';
+            ctx.shadowBlur = 10;
+            ctx.shadowColor = '#67e8f9';
+            ctx.fill();
+            ctx.shadowBlur = 0;
+
+            // Lantern Stalk & Huge Bulb
+            const lanternX = 14 * S; const lanternY = -15 * S;
+            ctx.beginPath(); ctx.moveTo(3*S, -11*S); 
+            ctx.quadraticCurveTo(8*S, -18*S, lanternX, lanternY);
+            ctx.strokeStyle = '#312e81'; ctx.lineWidth = 2; ctx.stroke(); // thicker stalk
+
+            // Bioluminescent pulse (cute soft glow)
+            const bulbPhase = Math.sin(t * 3.5) * 0.5 + 0.5;
+            ctx.beginPath(); ctx.arc(lanternX, lanternY, 2.5*S, 0, Math.PI*2); // big round bulb
+            ctx.fillStyle = '#cffafe'; ctx.shadowBlur = 24 + bulbPhase * 12; ctx.shadowColor = `rgba(103, 232, 249, ${0.7 + bulbPhase*0.3})`; ctx.fill();
+            // Inner hot core
+            ctx.beginPath(); ctx.arc(lanternX, lanternY, 1.2*S, 0, Math.PI*2);
+            ctx.fillStyle = '#ffffff'; ctx.fill(); ctx.shadowBlur = 0;
+
+            ctx.restore();
+        };
+
+        const drawFish = (t: number, mouthOpen: boolean, depthPhase: number) => {
+            // Keep one silhouette at a time. Only the colour/light changes
+            // through the middle so the two complete fish never overlap.
+            const clamp01 = (value: number) => Math.max(0, Math.min(1, value));
+            const deepeningPhase = clamp01((depthPhase - 0.08) / 0.72);
+            const anglerfishStart = 0.9;
+
+            if (depthPhase < anglerfishStart) {
+                // A faint lamp grows in behind the clownfish as it gets deeper.
+                if (deepeningPhase > 0) {
+                    ctx.save();
+                    ctx.globalAlpha = 0.18 * deepeningPhase;
+                    const glow = ctx.createRadialGradient(19, -20, 0, 19, -20, 11);
+                    glow.addColorStop(0, 'rgba(207,250,254,0.95)');
+                    glow.addColorStop(1, 'rgba(103,232,249,0)');
+                    ctx.fillStyle = glow;
+                    ctx.beginPath(); ctx.arc(19, -20, 11, 0, Math.PI * 2); ctx.fill();
+
+                    ctx.globalAlpha = 0.28 * deepeningPhase;
+                    ctx.strokeStyle = '#312e81'; ctx.lineWidth = 1.2;
+                    ctx.beginPath(); ctx.moveTo(4, -11); ctx.quadraticCurveTo(10, -17, 19, -20); ctx.stroke();
+                    ctx.restore();
+                }
+
+                drawClownfish(t, mouthOpen);
+
+                // Gradually shift the existing fish towards a darker blue-purple
+                // without changing its shape or drawing a second fish over it.
+                if (deepeningPhase > 0) {
+                    ctx.save();
+                    ctx.globalCompositeOperation = 'source-atop';
+                    ctx.fillStyle = `rgba(45, 76, 164, ${0.22 * deepeningPhase})`;
+                    ctx.fillRect(-38, -24, 66, 48);
+                    ctx.fillStyle = `rgba(5, 12, 36, ${0.10 * deepeningPhase})`;
+                    ctx.fillRect(-38, -24, 66, 48);
+                    ctx.restore();
+                }
+            } else {
+                // Switch only once the cursor is already deep, keeping the
+                // anglerfish drawing itself unchanged.
+                drawAnglerfish(t, mouthOpen);
+            }
+        };
+
+        // ── Animation loop ───────────────────────────────────────────────────
+        let last = performance.now();
+
+        const LERP_POS   = 0.28;   // position follow speed
+        const LERP_PITCH = 0.09;   // pitch (up/down tilt) smoothing
+        const VEL_DECAY  = 0.82;   // velocity decay per frame
+        // Hysteresis thresholds for left/right flip — prevents flickering at zero
+        const FLIP_TO_LEFT_VX  = -1.2;  // flip to face-left when vx drops below this
+        const FLIP_TO_RIGHT_VX =  1.2;  // flip to face-right when vx rises above this
+
+        const loop = (now: number) => {
+            s.raf = requestAnimationFrame(loop);
+            const dt = Math.min((now - last) / 16.67, 4); // normalised to 60 fps
+            last = now;
+            const t = now * 0.001;
+
+            // Smooth position
+            const prevX = s.x, prevY = s.y;
+            s.x += (s.rawX - s.x) * LERP_POS * dt;
+            s.y += (s.rawY - s.y) * LERP_POS * dt;
+
+            // Velocity (pixels / normalised-frame at 60 fps)
+            const frameVX = (s.x - prevX);
+            const frameVY = (s.y - prevY);
+            s.vx = s.vx * VEL_DECAY + frameVX * (1 - VEL_DECAY);
+            s.vy = s.vy * VEL_DECAY + frameVY * (1 - VEL_DECAY);
+
+            // Advance swim phase proportional to speed
+            const speed = Math.hypot(s.vx, s.vy);
+            s.phase += speed * 0.18 * dt;
+
+            // ── Facing direction (hysteresis flip) ──────────────────────────
+            // Fish always stays right-side-up; we mirror it with scale(-1,1)
+            // to face left instead of rotating 180° (which would flip the dorsal fin).
+            if (s.vx < FLIP_TO_LEFT_VX)  s.facingLeft = true;
+            if (s.vx > FLIP_TO_RIGHT_VX) s.facingLeft = false;
+
+            // ── Pitch angle: up/down tilt ───────────────────────────────────
+            // Compute target pitch relative to the fish's current facing direction
+            // so tilting up/down works correctly regardless of left/right.
+            let targetPitch = 0;
+            if (speed > 0.4) {
+                const signedVx = s.facingLeft ? -s.vx : s.vx; // always positive when moving forward
+                targetPitch = Math.atan2(s.vy, signedVx + 0.01) * 0.72; // dampen full tilt
+            }
+            // Smooth pitch
+            let dp = targetPitch - s.pitchAngle;
+            while (dp >  Math.PI) dp -= Math.PI * 2;
+            while (dp < -Math.PI) dp += Math.PI * 2;
+            s.pitchAngle += dp * LERP_PITCH * dt;
+
+            // Calculate depthPhase (fade into Anglerfish at the bottom 35% of the screen)
+            const logicalHeight = canvas.height / dpr;
+            const deepStart = logicalHeight * 0.65;
+            const deepEnd = logicalHeight * 0.85;
+            let depthPhase = 0;
+            if (s.y > deepStart) {
+                depthPhase = Math.min(1.0, (s.y - deepStart) / (deepEnd - deepStart));
+            }
+
+            // Clear
+            ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+            // Draw
+            ctx.save();
+            ctx.scale(dpr, dpr);
+            ctx.translate(s.x, s.y);
+            // Apply facing direction THEN pitch so dorsal fin stays on top always
+            if (s.facingLeft) ctx.scale(-1, 1);
+            ctx.rotate(s.pitchAngle);
+            drawFish(t, s.mouthOpen, depthPhase);
+            ctx.restore();
+        };
+
+        s.raf = requestAnimationFrame(loop);
+        return () => {
+            cancelAnimationFrame(s.raf);
+            window.removeEventListener('resize', resize);
+        };
+    }, []);
+
+    // Fine pointer only — on touch devices just render nothing (canvas stays invisible)
+    return (
+        <canvas
+            ref={canvasRef}
+            className="fixed inset-0 z-[9999] pointer-events-none"
+            aria-hidden="true"
+        />
+    );
+}
+
+const bubbleVertexShader = `
+uniform float uTime;
+attribute float scale;
+attribute float speed;
+attribute float seed;
+varying float vAlpha;
+
+void main() {
+    vec3 pos = position;
+    float rise = mod(uTime * speed + seed * 16.0, 24.0) - 12.0;
+    pos.y += rise;
+    pos.x += sin(uTime * 0.65 + seed * 7.0) * 0.22;
+
+    vec4 mvPosition = modelViewMatrix * vec4(pos, 1.0);
+    vAlpha = smoothstep(-22.0, -5.0, mvPosition.z) * (1.0 - smoothstep(-2.0, 2.0, mvPosition.z));
+
+    gl_PointSize = scale * (120.0 / max(0.1, -mvPosition.z));
+    gl_Position = projectionMatrix * mvPosition;
+}
+`;
+
+const bubbleFragmentShader = `
+varying float vAlpha;
+
+void main() {
+    vec2 uv = gl_PointCoord - vec2(0.5);
+    float r = length(uv);
+    if (r > 0.5) discard;
+
+    float ring = smoothstep(0.5, 0.34, r) - smoothstep(0.34, 0.2, r);
+    float core = smoothstep(0.34, 0.0, r);
+    float alpha = (ring * 0.50 + core * 0.12) * vAlpha;
+
+    vec3 col = mix(vec3(0.70, 0.94, 0.98), vec3(0.94, 1.0, 1.0), ring);
+    gl_FragColor = vec4(col, clamp(alpha, 0.0, 0.30));
+}
+`;
+
+function OceanSurface() {
+    const meshRef = useRef<THREE.Mesh>(null);
+    const uniforms = useMemo(() => ({
+        uTime: { value: 0 },
+        uColorDeep: { value: new THREE.Color('#0b3a49') },
+        uColorShallow: { value: new THREE.Color('#5acfd9') },
+        uResolution: { value: new THREE.Vector2(1, 1) },
+        uRipples: { value: new Float32Array(30) }
+    }), []);
+    const resolution = uniforms.uResolution.value;
+
+    const ripplesRef = useRef<{ x: number; y: number; time: number }[]>(
+        Array.from({ length: 10 }, () => ({ x: -1, y: -1, time: 999.0 }))
+    );
+    const rippleIdx = useRef(0);
+    const lastPos = useRef({ x: -999, y: -999 });
+
+    useEffect(() => {
+        const updateResolution = () => {
+            resolution.set(window.innerWidth, window.innerHeight);
+        };
+        updateResolution();
+        window.addEventListener('resize', updateResolution);
+
+        const addRipple = (clientX: number, clientY: number) => {
+            const x = clientX / window.innerWidth;
+            const y = 1.0 - (clientY / window.innerHeight);
+            ripplesRef.current[rippleIdx.current] = { x, y, time: 0.0 };
+            rippleIdx.current = (rippleIdx.current + 1) % 10;
+        };
+
+        const onPointerDown = (e: PointerEvent) => {
+            if (e.pointerType === 'touch') return; // Handled by touchstart for better mobile support
+            lastPos.current = { x: e.clientX, y: e.clientY };
+            addRipple(e.clientX, e.clientY);
+        };
+
+        const onTouchStart = (e: TouchEvent) => {
+            for (let i = 0; i < e.changedTouches.length; i++) {
+                const touch = e.changedTouches[i];
+                lastPos.current = { x: touch.clientX, y: touch.clientY };
+                addRipple(touch.clientX, touch.clientY);
+            }
+        };
+
+        // Add event listeners to 'document' rather than 'window', and use { capture: true } 
+        // to intercept events from the top-down BEFORE any PWA/iOS bubbling issues swallow them.
+        document.addEventListener('pointerdown', onPointerDown, { capture: true, passive: true });
+        document.addEventListener('touchstart', onTouchStart, { capture: true, passive: true });
+
+        return () => {
+            window.removeEventListener('resize', updateResolution);
+            document.removeEventListener('pointerdown', onPointerDown, { capture: true });
+            document.removeEventListener('touchstart', onTouchStart, { capture: true });
+        };
+    }, [resolution]);
+
+    useFrame((state, delta) => {
+        // R3F uniforms are mutable GPU inputs, not React state.
+        // eslint-disable-next-line react-hooks/immutability
+        uniforms.uTime.value = state.clock.getElapsedTime() * 0.85;
+
+        const flatRipples = uniforms.uRipples.value;
+        for (let i = 0; i < 10; i++) {
+            const r = ripplesRef.current[i];
+            r.time += delta;
+
+            flatRipples[i * 3 + 0] = r.x;
+            flatRipples[i * 3 + 1] = r.y;
+            flatRipples[i * 3 + 2] = r.time;
+        }
+    });
+
+    return (
+        <mesh ref={meshRef} position={[0, 0, -1]}>
+            <planeGeometry args={[100, 100]} />
+            <shaderMaterial
+                vertexShader={waveVertexShader}
+                fragmentShader={waveFragmentShader}
+                uniforms={uniforms}
+                transparent={true}
+                depthWrite={false}
+            />
+        </mesh>
+    );
+}
+
+function OceanBubbles() {
+    const uniforms = useMemo(() => ({ uTime: { value: 0 } }), []);
+    const { positions, scales, speeds, seeds } = OCEAN_BUBBLE_DATA;
+
+    useFrame((state) => {
+        // R3F uniforms are mutable GPU inputs, not React state.
+        // eslint-disable-next-line react-hooks/immutability
+        uniforms.uTime.value = state.clock.getElapsedTime();
+    });
+
+    return (
+        <points>
+            <bufferGeometry>
+                <bufferAttribute attach="attributes-position" args={[positions, 3]} />
+                <bufferAttribute attach="attributes-scale" args={[scales, 1]} />
+                <bufferAttribute attach="attributes-speed" args={[speeds, 1]} />
+                <bufferAttribute attach="attributes-seed" args={[seeds, 1]} />
+            </bufferGeometry>
+            <shaderMaterial
+                vertexShader={bubbleVertexShader}
+                fragmentShader={bubbleFragmentShader}
+                uniforms={uniforms}
+                transparent={true}
+                depthWrite={false}
+                blending={THREE.AdditiveBlending}
+            />
+        </points>
+    );
+}
+
+function MarineSnow() {
+    const reducedMotion = useReducedMotion();
+
+    return (
+        <div data-testid="marine-snow" className="fixed inset-0 pointer-events-none z-2 overflow-hidden" aria-hidden="true">
+            {marineSnowSpecs.map((particle, index) => (
+                <motion.span
+                    key={`marine-snow-${index}`}
+                    className="absolute rounded-full bg-cyan-50/75"
+                    style={{
+                        left: `${particle.left}%`,
+                        top: `${(index * 19) % 112 - 12}%`,
+                        width: particle.size,
+                        height: particle.size,
+                        opacity: particle.opacity,
+                        filter: particle.blur ? `blur(${particle.blur}px)` : undefined,
+                        boxShadow: '0 0 8px rgba(191,235,255,0.28)',
+                    }}
+                    animate={reducedMotion
+                        ? { x: 0, y: 0, opacity: particle.opacity }
+                        : { x: [0, particle.drift, particle.drift * -0.55, 0], y: ['-8vh', '38vh', '78vh', '116vh'], opacity: [0, particle.opacity, particle.opacity * 0.72, 0] }}
+                    transition={reducedMotion
+                        ? { duration: 0 }
+                        : { duration: particle.duration, delay: particle.delay, repeat: Infinity, ease: 'linear' }}
+                />
+            ))}
+        </div>
+    );
+}
+
+function OceanRainEffects() {
+    return (
+        <div className="fixed inset-0 pointer-events-none z-2 overflow-hidden" aria-hidden="true">
+            {rainLightSpecs.map((light, index) => (
+                <motion.div
+                    key={`rain-light-column-${index}`}
+                    className="absolute -top-[12%] h-[84%] rounded-full blur-2xl mix-blend-screen"
+                    style={{
+                        left: light.left,
+                        width: `${light.width}%`,
+                        transform: `rotate(${light.rotate}deg)`,
+                        background: 'linear-gradient(to bottom, rgba(95,196,220,0) 0%, rgba(95,196,220,0.08) 45%, rgba(147,205,255,0.025) 72%, rgba(95,196,220,0) 100%)',
+                    }}
+                    animate={{ opacity: [0.10, 0.28, 0.10], x: ['-2%', '2%', '-2%'] }}
+                    transition={{ duration: light.duration, delay: light.delay, repeat: Infinity, ease: 'easeInOut' }}
+                />
+            ))}
+        </div>
+    );
+}
+
+function OceanBackdrop({ weather }: { weather: WeatherType }) {
+    const isRain = weather === 'Rain';
+    const isSnow = weather === 'Snow';
+    const webglSupported = useWebGLSupport();
+    return (
+        <>
+            <div className="fixed inset-0 pointer-events-none z-0 bg-[#041116]" />
+
+            {webglSupported && (
+                <div className="fixed inset-0 pointer-events-none z-1">
+                    <Canvas camera={{ position: [0, 0, 0], fov: 75 }}>
+                        <color attach="background" args={['#041116']} />
+                        <OceanSurface />
+                        <OceanBubbles />
+                    </Canvas>
+                </div>
+            )}
+
+            <div
+                className="fixed inset-0 pointer-events-none z-2"
+                style={{
+                    background: 'radial-gradient(circle at 20% 18%, rgba(120, 247, 225, 0.10) 0%, rgba(120, 247, 225, 0.02) 20%, transparent 42%), radial-gradient(circle at 78% 20%, rgba(96, 165, 250, 0.12) 0%, rgba(96, 165, 250, 0.02) 24%, transparent 46%), linear-gradient(180deg, rgba(7,33,42,0.35) 0%, rgba(4,17,22,0.28) 45%, rgba(2,10,14,0.62) 100%)',
+                }}
+            />
+
+            <motion.div
+                className="fixed inset-0 pointer-events-none z-2 opacity-40"
+                animate={{ opacity: [0.24, 0.46, 0.24], x: ['-3%', '2%', '-3%'] }}
+                transition={{ duration: 13, repeat: Infinity, ease: 'easeInOut' }}
+                style={{
+                    background: 'radial-gradient(ellipse 44% 26% at 18% 28%, rgba(170, 246, 255, 0.20) 0%, rgba(170, 246, 255, 0.06) 40%, rgba(170,246,255,0.0) 100%), radial-gradient(ellipse 40% 24% at 74% 36%, rgba(137, 227, 255, 0.18) 0%, rgba(137, 227, 255, 0.05) 42%, rgba(137,227,255,0.0) 100%)',
+                }}
+            />
+
+            <motion.div
+                className="fixed inset-0 pointer-events-none z-2 opacity-50"
+                animate={{ backgroundPosition: ['0px 0px', '96px -52px', '0px 0px'] }}
+                transition={{ duration: 16, repeat: Infinity, ease: 'linear' }}
+                style={{
+                    background: 'repeating-linear-gradient(165deg, rgba(255,255,255,0.00) 0px, rgba(255,255,255,0.00) 34px, rgba(135, 245, 229, 0.035) 35px, rgba(135, 245, 229, 0.00) 64px)',
+                }}
+            />
+
+            {bubbleSpecs.map((bubble, index) => (
+                <motion.div
+                    key={index}
+                    className={`fixed pointer-events-none z-2 rounded-full bg-white/5 backdrop-blur-sm ${isRain ? 'border border-blue-100/35' : 'border border-white/20'}`}
+                    style={{
+                        left: bubble.left,
+                        top: bubble.top,
+                        width: bubble.size,
+                        height: bubble.size,
+                        boxShadow: isRain
+                            ? 'inset 0 0 18px rgba(191,219,254,0.17), 0 0 24px rgba(96,165,250,0.12)'
+                            : 'inset 0 0 18px rgba(255,255,255,0.12), 0 0 24px rgba(94,234,212,0.08)',
+                    }}
+                    animate={{
+                        y: [-6, -34, -6],
+                        x: [0, 6, -4, 0],
+                        opacity: [0.22, 0.48, 0.18],
+                    }}
+                    transition={{
+                        duration: bubble.duration,
+                        repeat: Infinity,
+                        ease: 'easeInOut',
+                        delay: bubble.delay,
+                    }}
+                />
+            ))}
+
+            {isRain && <OceanRainEffects />}
+            {isSnow && <MarineSnow />}
+
+            {fishSpecs.map((fish, index) => (
+                <motion.div
+                    key={index}
+                    className="fixed pointer-events-none z-2"
+                    style={{ left: fish.left, top: fish.top, width: fish.width }}
+                    animate={{ x: [0, 36, 0], y: [0, -10, 0], opacity: [0.10, 0.18, 0.10] }}
+                    transition={{ duration: fish.duration, repeat: Infinity, ease: 'easeInOut', delay: fish.delay }}
+                >
+                    <div className="relative h-6 w-full">
+                        <div className="absolute left-0 top-1/2 h-4 -translate-y-1/2 rounded-full bg-linear-to-r from-teal-200/0 via-teal-200/20 to-cyan-100/10" style={{ width: fish.width }} />
+                        <div className="absolute right-0 top-1/2 h-0 w-0 -translate-y-1/2 border-y-8 border-y-transparent border-l-14 border-l-cyan-100/10" />
+                    </div>
+                </motion.div>
+            ))}
+
+            <div
+                className="fixed inset-x-0 bottom-0 h-[32vh] pointer-events-none z-2"
+                style={{
+                    background: 'linear-gradient(to top, rgba(2,10,14,0.94) 0%, rgba(4,17,22,0.74) 42%, rgba(4,17,22,0) 100%)',
+                }}
+            />
+        </>
+    );
+}
+
+export { FishCursor, OceanBackdrop };
